+++ src/pipes/index.ts
@@ -1,5 +1,6 @@
 export * from './cellContext/cellContext.pipe';
 export * from './cssGridTemplateColumns/cssGridTemplateColumns.pipe';
 export * from './dataCellContext/dataCellContext.pipe';
+export * from './infinityNaN/infinityNaN.pipe';
 export * from './readValue/readValue.pipe';
 export * from './resolveForwardRef/resolveForwardRef.pipe';