+++ src/plugins/dataLoader/async/asyncDataLoader.component.ts
@@ -69,16 +69,9 @@ export class AsyncDataLoaderSAComponent<TData = unknown, TOrdering = unknown> ex
 
         this.ɵstate.set((this.ɵresult && this.ɵresult().data && this.ɵresult().data.length) ? DataLoaderState.DataLoading : DataLoaderState.NoDataLoading);
 
-        const result = await this.ɵoptions.dataCallback(this.paging?.page ?? 1, this.paging?.itemsPerPage ?? 0, this.ordering?.ordering());
+        const result = await this.ɵoptions.dataCallback(this.paging?.page() ?? 1, this.paging?.itemsPerPage() ?? 0, this.ordering?.ordering());
 
         this.ɵstate.set((result && result.data && result.data.length) ? DataLoaderState.Loaded : DataLoaderState.NoData);
-
-        if(this.paging)
-        {
-            this.paging.totalCount = result.totalCount;
-            this.paging.invalidateVisuals();
-        }
-
         this.ɵresult.set(result);
     }
 }