+++ src/plugins/dataLoader/sync/syncDataLoader.component.ts
@@ -104,8 +104,8 @@ export class SyncDataLoaderSAComponent<TData = unknown, TOrdering = unknown> ext
         }
 
         data = await lastValueFrom(from(data)
-            .pipe(skip(((this.paging?.page ?? 1) - 1) * (isNaN(this.paging?.itemsPerPage ?? 0) ? 0 : (this.paging?.itemsPerPage ?? 0))),
-                  isNaN(this.paging?.itemsPerPage ?? 0) ? ((source: Observable<TData>) => source) : take(this.paging?.itemsPerPage ?? 0),
+            .pipe(skip(((this.paging?.page() ?? 1) - 1) * (isNaN(this.paging?.itemsPerPage() ?? 0) ? 0 : (this.paging?.itemsPerPage() ?? 0))),
+                  isNaN(this.paging?.itemsPerPage() ?? 0) ? ((source: Observable<TData>) => source) : take(this.paging?.itemsPerPage() ?? 0),
                   toArray()));
 
         this.ɵstate.set((data && data.length) ? DataLoaderState.Loaded : DataLoaderState.NoData);