+++ src/plugins/paging/no/noPaging.component.ts
@@ -1,9 +1,8 @@
-import {ChangeDetectionStrategy, Component, ElementRef, ChangeDetectorRef, Optional, Inject} from '@angular/core';
+import {ChangeDetectionStrategy, Component, ValueProvider} from '@angular/core';
 
 import {PagingAbstractComponent} from '../pagingAbstract.component';
 import {NoPagingOptions, NoPaging} from './noPaging.interface';
-import {GRID_PLUGIN_INSTANCES, PAGING_OPTIONS} from '../../../misc/tokens';
-import {GridPluginInstances} from '../../../interfaces';
+import {DEFAULT_OPTIONS} from '../../../misc/tokens';
 
 /**
  * Default options for paging
@@ -23,9 +22,17 @@ const defaultOptions: NoPagingOptions =
     selector: 'ng-no-paging',
     template: '',
     standalone: true,
+    providers:
+    [
+        <ValueProvider>
+        {
+            provide: DEFAULT_OPTIONS,
+            useValue: defaultOptions,
+        },
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class NoPagingSAComponent extends PagingAbstractComponent<unknown, NoPagingOptions> implements NoPaging
+export class NoPagingSAComponent extends PagingAbstractComponent<unknown, NoPagingOptions> implements NoPaging<NoPagingOptions>
 {
     //######################### public properties - implementation of NoPaging #########################
 
@@ -33,28 +40,4 @@ export class NoPagingSAComponent extends PagingAbstractComponent<unknown, NoPagi
      * @inheritdoc
      */
     public firstItemIndex: number = 0;
-
-    /**
-     * @inheritdoc
-     */
-    public page: number = 1;
-
-    /**
-     * @inheritdoc
-     */
-    public itemsPerPage: number = NaN;
-
-    /**
-     * @inheritdoc
-     */
-    public totalCount: number = 0;
-
-    //######################### constructor #########################
-    constructor(pluginElement: ElementRef,
-                changeDetector: ChangeDetectorRef,
-                @Inject(GRID_PLUGIN_INSTANCES) @Optional() gridPlugins: GridPluginInstances|undefined|null,
-                @Inject(PAGING_OPTIONS) @Optional() options?: NoPagingOptions,)
-    {
-        super(pluginElement, changeDetector, gridPlugins, defaultOptions, options);
-    }
 }