+++ src/plugins/paging/no/noPaging.interface.ts
@@ -3,13 +3,13 @@ import {Paging, PagingOptions} from '../../../interfaces';
 /**
  * No paging options for paging
  */
-export interface NoPagingOptions<TCssClasses = unknown> extends PagingOptions<TCssClasses>
+export interface NoPagingOptions extends PagingOptions<unknown>
 {
 }
 
 /**
  * Public API for no paging
  */
-export interface NoPaging extends Paging
+export interface NoPaging<TOptions extends NoPagingOptions = NoPagingOptions> extends Paging<TOptions>
 {
 }